+++ app/javascript/components/game/map/MapCounterOverlay.tsx
@@ -227,7 +227,8 @@ export default function MapCounterOverlay({
           const dblwidth = outwidth*2
           const x = layout.x + i*(160+dblwidth) + dblwidth*2
           let thisButtons = false
-          if (counter.hasUnit && (counter.unit.selected || counter.unit.targetSelected) && yy) {
+          if ((counter.hasUnit || (counter.hasFeature && map.game?.phase === gamePhaseType.Deploy)) &&
+              (counter.targetUF.selected || counter.targetUF.targetSelected) && yy) {
             const controls = counterActionButtons(map, x, layout.y2 - outwidth + 2, maxY, counter)
             for (const c of controls) {
               thisButtons = true