+++ app/javascript/engine/control/reactionFire.ts
@@ -63,7 +63,7 @@ export function reactionAvailableCoords(game: Game): Coordinate[] {
           if (c.unit.areaFire || c.unit.isBroken) { continue }
           if (c.unit.operated) {
             if (c.unit.parent === undefined) { continue }
-            if (c.unit.parent.isBroken || c.unit.parent.isExhausted) { continue }
+            if (c.unit.parent.isBroken || c.unit.parent.isExhausted || c.unit.parent.pinned) { continue }
           }
           if (c.unit.isWreck) { continue }
           if ((c.unit.canCarrySupport || c.unit.rapidFire) && otherUnit?.armored) { continue }