+++ app/models/game_action.rb
@@ -16,13 +16,14 @@ class GameAction < ApplicationRecord
 
   def self.create_action(params)
     game = Game.find_by(id: params[:game_id])
-    user_id = params[:user_id]
+    user = User.find_by(id: params[:user_id])
     # Sometimes the moves get recorded out-of-order when zooming through unused
     # phases, and it can happen at the end:
     return nil unless game&.in_progress? || game&.complete?
-    return nil if game.player_one.id != user_id.to_i && game.player_two.id != user_id.to_i
+    return nil if game.player_one.id != user&.id && game.player_two.id != user&.id
 
     params[:data] = JSON.parse(params[:data])
+    user.update!(notified: false)
 
     GameAction.create!(params)
   end