+++ config/routes.rb
@@ -13,8 +13,10 @@ Rails.application.routes.draw do
           post "check_conflict"
           post "set_recovery"
           post "password_reset"
+          post "notify"
           put "toggle_dev"
           put "toggle_banned"
+          put "toggle_notifications"
           get "stats"
           get "all"
         end