+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2026_04_19_003510) do
+ActiveRecord::Schema[7.0].define(version: 2026_04_21_210121) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -98,6 +98,8 @@ ActiveRecord::Schema[7.0].define(version: 2026_04_19_003510) do
     t.boolean "developer", default: false, null: false
     t.boolean "admin", default: false, null: false
     t.boolean "banned", default: false, null: false
+    t.boolean "notifications", default: true, null: false
+    t.boolean "notified", default: false, null: false
     t.index ["email"], name: "index_users_on_email"
     t.index ["id"], name: "index_users_on_id"
     t.index ["username"], name: "index_users_on_username"