+++ changelog.md
@@ -1,3 +1,11 @@
+## 0.62 Alpha
+
+- Finish email notifications / allow notification toggle
+- Allow undeploy of deployed features
+- Bug fix: pinned parents of operated vehicles should not trigger reaction fire
+- Delete (some) games, clean up scenarios
+- Update 603
+
 ## 0.61 Alpha
 
 - Major selection refactor to improve performance, make UX changes cleaner